update order model

+++ src/scripts/ddl.sql
@@ -178,7 +178,7 @@ CREATE TABLE "order" (
     user_mail TEXT NOT NULL REFERENCES users(mail),
     address_id INTEGER NOT NULL REFERENCES address(address_id),
 	status TEXT NOT NULL,
-    total NUMERIC(6,2) NOT NULL,
+    total FLOAT NOT NULL,
     order_date DATE NOT NULL,
     delivery_date DATE NOT NULL,
     shipping_fee INTEGER NOT NULL,
@@ -195,7 +195,7 @@ CREATE TABLE order_detail (
     order_id UUID NOT NULL REFERENCES "order"(order_id),
     product_id TEXT NOT NULL REFERENCES product(product_id),
 	quantity INTEGER NOT NULL,
-    price NUMERIC(5,2) NOT NULL,
+    price FLOAT NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
 	updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
